Refactor: Consolidate enhanced feature generators
