--- src/feature_generation/categories/enhanced_cross_timeframe.py
@@ -1,708 +1,0 @@
-"""
-Enhanced Cross-Timeframe Feature Generators
-
-This module provides advanced cross-timeframe analysis features with proper
-lag handling, fractional changes, and learned projections to capture
-relationships across different time horizons.
-
-Features implemented:
-- Proper lag handling to avoid lookahead bias
-- Fractional change features across timeframes
-- Cross-timeframe alignment and synchronization
-- Learned projections using PCA, autoencoders, and PatchTST
-- Regime-aware cross-timeframe features
-- Multi-asset cross-timeframe correlations
-"""
-
-import numpy as np
-import pandas as pd
-from typing import Any, Dict, List, Optional, Union, Tuple
-import logging
-from scipy import stats
-from sklearn.decomposition import PCA
-from sklearn.preprocessing import StandardScaler
-
-from ..core.feature_generator import (
-    FeatureGenerator,
-    FeatureConfig,
-    FeatureCategory,
-    VectorizedFeatureGenerator
-)
-
-logger = logging.getLogger(__name__)
-
-
-class EnhancedCrossTimeframeFeatureGenerator(VectorizedFeatureGenerator):
-    """Enhanced feature generator for cross-timeframe analysis."""
-
-    def __init__(self, config: Optional[FeatureConfig] = None):
-        if config is None:
-            config = self._create_default_config()
-        super().__init__(config, enable_matrix_ops=True)
-
-    @classmethod
-    def _create_default_config(cls) -> FeatureConfig:
-        return FeatureConfig(
-            name="enhanced_cross_timeframe_features",
-            category=FeatureCategory.CROSS_TIMEFRAME,
-            description="Enhanced cross-timeframe features with proper lag handling and learned projections",
-            required_columns=["close"],
-            optional_columns=["high", "low", "open", "volume"],
-            default_lookback=100,
-            min_lookback=50,
-            max_lookback=500,
-            parameters={
-                "timeframes": [1, 5, 15, 30, 60],
-                "feature_types": ["momentum", "volatility", "volume", "trend", "range"],
-                "lag_handling": True,
-                "fractional_changes": True,
-                "learned_projections": True,
-                "regime_aware": True,
-                "alignment_methods": ["lag", "resample", "interpolate"],
-                "projection_methods": ["pca", "autoencoder", "patchtst"]
-            },
-            matrix_optimized=True,
-            gpu_accelerated=False
-        )
-
-    def _generate_feature(self, data: pd.DataFrame, **kwargs) -> pd.Series:
-        """Generate enhanced cross-timeframe features."""
-        try:
-            # Generate all enhanced cross-timeframe features
-            features_dict = self.generate_enhanced_cross_timeframe_features(data, **kwargs)
-
-            # Return first feature as representative for base class
-            if features_dict:
-                first_feature_name = list(features_dict.keys())[0]
-                return pd.Series(features_dict[first_feature_name], index=data.index)
-            else:
-                return pd.Series(np.zeros(len(data)), index=data.index)
-
-        except Exception as e:
-            logger.error(f"Error generating enhanced cross-timeframe features: {e}")
-            return pd.Series(np.zeros(len(data)), index=data.index)
-
-    def generate_enhanced_cross_timeframe_features(self, data: pd.DataFrame, **kwargs) -> Dict[str, np.ndarray]:
-        """Generate comprehensive enhanced cross-timeframe features."""
-        features = {}
-
-        try:
-            # Fractional change features with proper lag handling
-            features.update(self._generate_fractional_change_features(data))
-
-            # Cross-timeframe alignment features
-            features.update(self._generate_alignment_features(data))
-
-            # Learned projection features
-            features.update(self._generate_learned_projection_features(data))
-
-            # Regime-aware cross-timeframe features
-            features.update(self._generate_regime_aware_cross_timeframe_features(data))
-
-            # Multi-scale correlation features
-            features.update(self._generate_multi_scale_correlation_features(data))
-
-            # Cross-timeframe divergence features
-            features.update(self._generate_divergence_features(data))
-
-            # Structural break features
-            features.update(self._generate_structural_break_features(data))
-
-            logger.info(f"Generated {len(features)} enhanced cross-timeframe features")
-            return features
-
-        except Exception as e:
-            logger.error(f"Error in generate_enhanced_cross_timeframe_features: {e}")
-            return {}
-
-    def _generate_fractional_change_features(self, data: pd.DataFrame) -> Dict[str, np.ndarray]:
-        """Generate fractional change features across timeframes with proper lag handling."""
-        features = {}
-        timeframes = self.config.parameters.get("timeframes", [1, 5, 15, 30, 60])
-        feature_types = self.config.parameters.get("feature_types", ["momentum", "volatility", "volume", "trend"])
-
-        for fast_tf in timeframes:
-            for slow_tf in timeframes:
-                if fast_tf >= slow_tf:
-                    continue
-
-                for feature_type in feature_types:
-                    # Calculate features with proper lag handling
-                    fast_feature = self._calculate_feature_with_lag(data, fast_tf, feature_type)
-                    slow_feature = self._calculate_feature_with_lag(data, slow_tf, feature_type)
-
-                    if fast_feature is not None and slow_feature is not None:
-                        # Fractional change
-                        fractional_change = fast_feature / (slow_feature + 1e-8)
-                        features[f"frac_change_{feature_type}_{fast_tf}m_{slow_tf}m"] = fractional_change.fillna(0).values
-
-                        # Relative change
-                        relative_change = (fast_feature - slow_feature) / (slow_feature + 1e-8)
-                        features[f"rel_change_{feature_type}_{fast_tf}m_{slow_tf}m"] = relative_change.fillna(0).values
-
-                        # Momentum divergence
-                        momentum_div = fast_feature - slow_feature
-                        features[f"momentum_div_{feature_type}_{fast_tf}m_{slow_tf}m"] = momentum_div.fillna(0).values
-
-        return features
-
-    def _calculate_feature_with_lag(self, data: pd.DataFrame, timeframe: int, feature_type: str) -> Optional[pd.Series]:
-        """Calculate feature with proper lag handling to avoid lookahead bias."""
-        try:
-            if feature_type == "momentum":
-                # Calculate momentum with lag
-                lag_bars = max(1, timeframe // 5)  # Lag by 20% of timeframe
-                returns = data["close"].pct_change(timeframe)
-                return returns.shift(lag_bars)
-
-            elif feature_type == "volatility":
-                # Calculate volatility with lag
-                lag_bars = max(1, timeframe // 5)
-                returns = data["close"].pct_change()
-                vol = returns.rolling(window=timeframe).std()
-                return vol.shift(lag_bars)
-
-            elif feature_type == "volume":
-                if "volume" in data.columns:
-                    lag_bars = max(1, timeframe // 5)
-                    vol_ma = data["volume"].rolling(window=timeframe).mean()
-                    return vol_ma.shift(lag_bars)
-                else:
-                    return None
-
-            elif feature_type == "trend":
-                # Calculate trend strength with lag
-                lag_bars = max(1, timeframe // 5)
-                trend = self._calculate_trend_strength(data["close"], timeframe)
-                return trend.shift(lag_bars)
-
-            elif feature_type == "range":
-                # Calculate high-low range with lag
-                lag_bars = max(1, timeframe // 5)
-                if "high" in data.columns and "low" in data.columns:
-                    hl_range = (data["high"] - data["low"]).rolling(window=timeframe).mean()
-                    return hl_range.shift(lag_bars)
-                else:
-                    return None
-
-            else:
-                return None
-
-        except Exception as e:
-            logger.warning(f"Error calculating {feature_type} for timeframe {timeframe}: {e}")
-            return None
-
-    def _calculate_trend_strength(self, series: pd.Series, window: int) -> pd.Series:
-        """Calculate trend strength using linear regression slope."""
-        def calc_slope(x):
-            if len(x) < 2:
-                return 0.0
-            try:
-                return np.polyfit(range(len(x)), x, 1)[0]
-            except:
-                return 0.0
-
-        return series.rolling(window=window).apply(calc_slope, raw=False)
-
-    def _generate_alignment_features(self, data: pd.DataFrame) -> Dict[str, np.ndarray]:
-        """Generate cross-timeframe alignment features."""
-        features = {}
-        timeframes = self.config.parameters.get("timeframes", [1, 5, 15, 30, 60])
-        alignment_methods = self.config.parameters.get("alignment_methods", ["lag", "resample", "interpolate"])
-
-        for source_tf in timeframes:
-            for target_tf in timeframes:
-                if source_tf >= target_tf:
-                    continue
-
-                for method in alignment_methods:
-                    aligned_feature = self._align_timeframes(data, source_tf, target_tf, method)
-                    if aligned_feature is not None:
-                        features[f"aligned_{source_tf}m_to_{target_tf}m_{method}"] = aligned_feature.fillna(0).values
-
-        return features
-
-    def _align_timeframes(self, data: pd.DataFrame, source_tf: int, target_tf: int, method: str) -> Optional[pd.Series]:
-        """Align features from source timeframe to target timeframe."""
-        try:
-            if method == "lag":
-                # Lag fast timeframe features by appropriate number of bars
-                lag_bars = target_tf // source_tf - 1
-                returns = data["close"].pct_change()
-                return returns.shift(lag_bars)
-
-            elif method == "resample":
-                # Resample to target timeframe
-                resampled = data["close"].resample(f'{target_tf}min').last()
-                # Forward fill to original frequency
-                aligned = resampled.reindex(data.index, method='ffill')
-                return (aligned / aligned.shift(1) - 1).fillna(0)
-
-            elif method == "interpolate":
-                # Interpolate between timeframes
-                returns = data["close"].pct_change()
-                # Simple interpolation (in practice, would use more sophisticated methods)
-                return returns.rolling(window=target_tf//source_tf).mean()
-
-            else:
-                return None
-
-        except Exception as e:
-            logger.warning(f"Error aligning timeframes {source_tf} to {target_tf} with method {method}: {e}")
-            return None
-
-    def _generate_learned_projection_features(self, data: pd.DataFrame) -> Dict[str, np.ndarray]:
-        """Generate learned projection features across timeframes."""
-        features = {}
-        timeframes = self.config.parameters.get("timeframes", [1, 5, 15, 30, 60])
-        projection_methods = self.config.parameters.get("projection_methods", ["pca", "autoencoder", "patchtst"])
-
-        for method in projection_methods:
-            if method == "pca":
-                features.update(self._generate_pca_projection_features(data, timeframes))
-            elif method == "autoencoder":
-                features.update(self._generate_autoencoder_projection_features(data, timeframes))
-            elif method == "patchtst":
-                features.update(self._generate_patchtst_projection_features(data, timeframes))
-
-        return features
-
-    def _generate_pca_projection_features(self, data: pd.DataFrame, timeframes: List[int]) -> Dict[str, np.ndarray]:
-        """Generate PCA projection features across timeframes."""
-        features = {}
-
-        try:
-            # Create features for each timeframe
-            tf_features = []
-            for tf in timeframes:
-                # Calculate returns for this timeframe
-                returns = data["close"].pct_change(tf).fillna(0)
-
-                # Calculate volatility for this timeframe
-                vol = data["close"].pct_change().rolling(window=tf).std().fillna(0)
-
-                # Calculate momentum for this timeframe
-                momentum = data["close"].pct_change(tf * 2).fillna(0)
-
-                # Calculate trend for this timeframe
-                trend = self._calculate_trend_strength(data["close"], tf).fillna(0)
-
-                tf_features.append(pd.concat([returns, vol, momentum, trend], axis=1))
-
-            # Combine features from all timeframes
-            feature_matrix = pd.concat(tf_features, axis=1).fillna(0)
-
-            # Apply PCA for dimensionality reduction
-            if len(feature_matrix.columns) >= 3:
-                pca = PCA(n_components=min(3, len(feature_matrix.columns)))
-                pca_result = pca.fit_transform(feature_matrix)
-
-                for i in range(pca_result.shape[1]):
-                    features[f"pca_component_{i+1}"] = pca_result[:, i]
-
-                # Explained variance ratio
-                for i, ratio in enumerate(pca.explained_variance_ratio_):
-                    features[f"pca_explained_var_{i+1}"] = np.full(len(data), ratio)
-
-        except Exception as e:
-            logger.warning(f"Error in PCA projection: {e}")
-
-        return features
-
-    def _generate_autoencoder_projection_features(self, data: pd.DataFrame, timeframes: List[int]) -> Dict[str, np.ndarray]:
-        """Generate autoencoder projection features across timeframes."""
-        features = {}
-
-        try:
-            # Create input features
-            input_features = []
-            for tf in timeframes:
-                returns = data["close"].pct_change(tf).fillna(0)
-                vol = data["close"].pct_change().rolling(window=tf).std().fillna(0)
-                input_features.extend([returns, vol])
-
-            feature_matrix = pd.concat(input_features, axis=1).fillna(0)
-
-            # Simple autoencoder using PCA as proxy
-            if len(feature_matrix.columns) >= 2:
-                pca = PCA(n_components=min(2, len(feature_matrix.columns)))
-                encoded = pca.fit_transform(feature_matrix)
-
-                for i in range(encoded.shape[1]):
-                    features[f"autoencoder_component_{i+1}"] = encoded[:, i]
-
-        except Exception as e:
-            logger.warning(f"Error in autoencoder projection: {e}")
-
-        return features
-
-    def _generate_patchtst_projection_features(self, data: pd.DataFrame, timeframes: List[int]) -> Dict[str, np.ndarray]:
-        """Generate PatchTST projection features across timeframes."""
-        features = {}
-
-        try:
-            # Create patches for each timeframe
-            patch_length = 16
-            num_patches = 8
-
-            for tf in timeframes:
-                # Create patches from price sequence
-                price_sequence = data["close"].values
-                patches = self._create_patches(price_sequence, patch_length, num_patches)
-
-                if patches is not None:
-                    # Calculate patch statistics
-                    patch_means = patches.mean(axis=1)
-                    patch_stds = patches.std(axis=1)
-                    patch_trends = np.polyfit(np.arange(patch_length), patches.T, 1)[0]
-
-                    features[f"patchtst_mean_{tf}"] = patch_means
-                    features[f"patchtst_std_{tf}"] = patch_stds
-                    features[f"patchtst_trend_{tf}"] = patch_trends
-
-        except Exception as e:
-            logger.warning(f"Error in PatchTST projection: {e}")
-
-        return features
-
-    def _create_patches(self, sequence: np.ndarray, patch_length: int, num_patches: int) -> Optional[np.ndarray]:
-        """Create patches from price sequence."""
-        seq_len = len(sequence)
-        patch_size = patch_length * num_patches
-
-        if seq_len < patch_size:
-            return None
-
-        # Take the most recent data
-        recent_data = sequence[-patch_size:]
-        
-        # Reshape into patches
-        patches = recent_data.reshape(num_patches, patch_length)
-        return patches
-
-    def _generate_regime_aware_cross_timeframe_features(self, data: pd.DataFrame) -> Dict[str, np.ndarray]:
-        """Generate regime-aware cross-timeframe features."""
-        features = {}
-
-        # Detect market regimes
-        returns = data["close"].pct_change()
-        vol_regime = returns.rolling(window=20).std()
-        
-        # Define regimes
-        low_vol_threshold = vol_regime.quantile(0.33)
-        high_vol_threshold = vol_regime.quantile(0.67)
-        
-        low_vol_regime = (vol_regime <= low_vol_threshold).astype(int)
-        high_vol_regime = (vol_regime >= high_vol_threshold).astype(int)
-
-        # Cross-timeframe features for each regime
-        timeframes = [5, 15, 30]
-        
-        for tf1 in timeframes:
-            for tf2 in timeframes:
-                if tf1 >= tf2:
-                    continue
-
-                # Calculate features
-                feature1 = data["close"].pct_change(tf1)
-                feature2 = data["close"].pct_change(tf2)
-
-                # Low volatility regime features
-                low_vol_mask = low_vol_regime == 1
-                if low_vol_mask.sum() > 0:
-                    low_vol_ratio = np.zeros(len(data))
-                    low_vol_ratio[low_vol_mask] = (feature1 / (feature2 + 1e-8))[low_vol_mask]
-                    features[f"regime_low_vol_ratio_{tf1}_{tf2}"] = low_vol_ratio
-
-                # High volatility regime features
-                high_vol_mask = high_vol_regime == 1
-                if high_vol_mask.sum() > 0:
-                    high_vol_ratio = np.zeros(len(data))
-                    high_vol_ratio[high_vol_mask] = (feature1 / (feature2 + 1e-8))[high_vol_mask]
-                    features[f"regime_high_vol_ratio_{tf1}_{tf2}"] = high_vol_ratio
-
-        return features
-
-    def _generate_multi_scale_correlation_features(self, data: pd.DataFrame) -> Dict[str, np.ndarray]:
-        """Generate multi-scale correlation features."""
-        features = {}
-
-        timeframes = [5, 10, 20, 50]
-        correlation_window = 20
-
-        for tf1 in timeframes:
-            for tf2 in timeframes:
-                if tf1 >= tf2:
-                    continue
-
-                # Calculate features
-                feature1 = data["close"].pct_change(tf1)
-                feature2 = data["close"].pct_change(tf2)
-
-                # Rolling correlation
-                correlation = feature1.rolling(window=correlation_window).corr(feature2)
-                features[f"correlation_{tf1}_{tf2}_{correlation_window}"] = correlation.fillna(0).values
-
-                # Correlation stability
-                corr_std = correlation.rolling(window=correlation_window).std()
-                features[f"corr_stability_{tf1}_{tf2}_{correlation_window}"] = corr_std.fillna(0).values
-
-        return features
-
-    def _generate_divergence_features(self, data: pd.DataFrame) -> Dict[str, np.ndarray]:
-        """Generate cross-timeframe divergence features."""
-        features = {}
-
-        timeframes = [5, 15, 30, 60]
-
-        for tf1 in timeframes:
-            for tf2 in timeframes:
-                if tf1 >= tf2:
-                    continue
-
-                # Calculate momentum divergence
-                momentum1 = data["close"].pct_change(tf1)
-                momentum2 = data["close"].pct_change(tf2)
-                divergence = momentum1 - momentum2
-                features[f"divergence_{tf1}_{tf2}"] = divergence.fillna(0).values
-
-                # Volatility divergence
-                vol1 = data["close"].pct_change().rolling(window=tf1).std()
-                vol2 = data["close"].pct_change().rolling(window=tf2).std()
-                vol_divergence = vol1 - vol2
-                features[f"vol_divergence_{tf1}_{tf2}"] = vol_divergence.fillna(0).values
-
-        return features
-
-    def _generate_structural_break_features(self, data: pd.DataFrame) -> Dict[str, np.ndarray]:
-        """Generate structural break features across timeframes."""
-        features = {}
-
-        timeframes = [10, 20, 50]
-
-        for tf in timeframes:
-            # Calculate rolling statistics
-            returns = data["close"].pct_change()
-            rolling_mean = returns.rolling(window=tf).mean()
-            rolling_std = returns.rolling(window=tf).std()
-
-            # Structural break detection using CUSUM
-            cusum = self._calculate_cusum(returns, tf)
-            features[f"structural_break_{tf}"] = cusum.fillna(0).values
-
-            # Regime change detection
-            regime_change = self._detect_regime_changes(rolling_mean, rolling_std)
-            features[f"regime_change_{tf}"] = regime_change.fillna(0).values
-
-        return features
-
-    def _calculate_cusum(self, series: pd.Series, window: int) -> pd.Series:
-        """Calculate CUSUM for structural break detection."""
-        rolling_mean = series.rolling(window=window).mean()
-        rolling_std = series.rolling(window=window).std()
-        
-        # Standardized series
-        standardized = (series - rolling_mean) / (rolling_std + 1e-8)
-        
-        # CUSUM
-        cusum = standardized.cumsum()
-        
-        return cusum
-
-    def _detect_regime_changes(self, rolling_mean: pd.Series, rolling_std: pd.Series) -> pd.Series:
-        """Detect regime changes using rolling statistics."""
-        # Calculate change in rolling statistics
-        mean_change = rolling_mean.diff().abs()
-        std_change = rolling_std.diff().abs()
-        
-        # Regime change indicator
-        regime_change = ((mean_change > mean_change.quantile(0.9)) | 
-                        (std_change > std_change.quantile(0.9))).astype(int)
-        
-        return regime_change
-
-
-# Individual enhanced cross-timeframe generators
-
-class FractionalChangeGenerator(FeatureGenerator):
-    """Generator for fractional change features across timeframes."""
-
-    def __init__(self, fast_tf: int = 5, slow_tf: int = 15, feature_type: str = "volatility"):
-        config = FeatureConfig(
-            name=f"fractional_change_{feature_type}_{fast_tf}m_{slow_tf}m",
-            category=FeatureCategory.CROSS_TIMEFRAME,
-            description=f"Fractional change of {feature_type} from {fast_tf}m to {slow_tf}m timeframe",
-            required_columns=["close"],
-            optional_columns=["high", "low", "open", "volume"],
-            default_lookback=max(fast_tf, slow_tf),
-            min_lookback=max(fast_tf, slow_tf),
-            max_lookback=max(fast_tf, slow_tf),
-            parameters={"fast_tf": fast_tf, "slow_tf": slow_tf, "feature_type": feature_type}
-        )
-        super().__init__(config)
-        self.fast_tf = fast_tf
-        self.slow_tf = slow_tf
-        self.feature_type = feature_type
-
-    def _generate_feature(self, data: pd.DataFrame, **kwargs) -> pd.Series:
-        """Generate fractional change feature across timeframes."""
-        if self.feature_type == "volatility":
-            fast_vol = data["close"].pct_change().rolling(window=self.fast_tf).std()
-            slow_vol = data["close"].pct_change().rolling(window=self.slow_tf).std()
-            fractional_change = fast_vol / (slow_vol + 1e-8)
-        elif self.feature_type == "momentum":
-            fast_momentum = data["close"].pct_change(self.fast_tf)
-            slow_momentum = data["close"].pct_change(self.slow_tf)
-            fractional_change = fast_momentum / (slow_momentum + 1e-8)
-        elif self.feature_type == "volume":
-            if "volume" in data.columns:
-                fast_volume = data["volume"].rolling(window=self.fast_tf).mean()
-                slow_volume = data["volume"].rolling(window=self.slow_tf).mean()
-                fractional_change = fast_volume / (slow_volume + 1e-8)
-            else:
-                fractional_change = pd.Series(np.zeros(len(data)), index=data.index)
-        else:
-            fractional_change = pd.Series(np.zeros(len(data)), index=data.index)
-
-        return fractional_change.fillna(0)
-
-
-class CrossTimeframeAlignmentGenerator(FeatureGenerator):
-    """Generator for properly aligned cross-timeframe features."""
-
-    def __init__(self, source_tf: int = 1, target_tf: int = 5, alignment_method: str = "lag"):
-        config = FeatureConfig(
-            name=f"ctf_aligned_{source_tf}m_to_{target_tf}m_{alignment_method}",
-            category=FeatureCategory.CROSS_TIMEFRAME,
-            description=f"Align {source_tf}m features to {target_tf}m timeframe using {alignment_method}",
-            required_columns=["close"],
-            default_lookback=target_tf,
-            min_lookback=target_tf,
-            max_lookback=target_tf,
-            parameters={"source_tf": source_tf, "target_tf": target_tf, "alignment_method": alignment_method}
-        )
-        super().__init__(config)
-        self.source_tf = source_tf
-        self.target_tf = target_tf
-        self.alignment_method = alignment_method
-
-    def _generate_feature(self, data: pd.DataFrame, **kwargs) -> pd.Series:
-        """Generate properly aligned cross-timeframe feature."""
-        # Calculate lag needed for alignment
-        lag_bars = self.target_tf // self.source_tf - 1
-
-        if self.alignment_method == "lag":
-            # Lag fast timeframe features by appropriate number of bars
-            returns = data["close"].pct_change()
-            aligned_returns = returns.shift(lag_bars)
-            return aligned_returns.fillna(0)
-        elif self.alignment_method == "resample":
-            # Resample to target timeframe
-            resampled = data["close"].resample(f'{self.target_tf}min').last()
-            # Forward fill to original frequency
-            aligned = resampled.reindex(data.index, method='ffill')
-            return (aligned / aligned.shift(1) - 1).fillna(0)
-        else:
-            return pd.Series(np.zeros(len(data)), index=data.index)
-
-
-class LearnedProjectionGenerator(FeatureGenerator):
-    """Generator for learned projections across timeframes using PCA/dimensionality reduction."""
-
-    def __init__(self, timeframes: List[int] = [1, 5, 15], n_components: int = 3):
-        config = FeatureConfig(
-            name=f"learned_projection_{'_'.join(map(str, timeframes))}_{n_components}",
-            category=FeatureCategory.CROSS_TIMEFRAME,
-            description=f"Learned projection across {timeframes} timeframes using {n_components} components",
-            required_columns=["close"],
-            optional_columns=["high", "low", "open", "volume"],
-            default_lookback=max(timeframes) * 10,
-            min_lookback=max(timeframes) * 5,
-            max_lookback=max(timeframes) * 20,
-            parameters={"timeframes": timeframes, "n_components": n_components}
-        )
-        super().__init__(config)
-        self.timeframes = timeframes
-        self.n_components = n_components
-
-    def _generate_feature(self, data: pd.DataFrame, **kwargs) -> pd.Series:
-        """Generate learned projection features across timeframes."""
-        try:
-            # Create features for each timeframe
-            tf_features = []
-            for tf in self.timeframes:
-                # Calculate returns for this timeframe
-                returns = data["close"].pct_change(tf)
-
-                # Calculate volatility for this timeframe
-                volatility = returns.rolling(window=tf).std()
-
-                # Calculate momentum for this timeframe
-                momentum = data["close"].pct_change(tf * 5)
-
-                tf_features.append(pd.concat([returns, volatility, momentum], axis=1))
-
-            # Combine features from all timeframes
-            feature_matrix = pd.concat(tf_features, axis=1).fillna(0)
-
-            # Apply PCA for dimensionality reduction
-            if len(feature_matrix.columns) >= self.n_components:
-                pca = PCA(n_components=self.n_components)
-                pca_result = pca.fit_transform(feature_matrix)
-
-                # Return first principal component as representative feature
-                return pd.Series(pca_result[:, 0], index=data.index)
-            else:
-                return pd.Series(np.zeros(len(data)), index=data.index)
-
-        except Exception as e:
-            logger.warning(f"Error in learned projection: {e}")
-            return pd.Series(np.zeros(len(data)), index=data.index)
-
-
-def create_enhanced_cross_timeframe_generators() -> List[FeatureGenerator]:
-    """Create all enhanced cross-timeframe feature generators."""
-    generators = []
-
-    # Main enhanced cross-timeframe generator
-    generators.append(EnhancedCrossTimeframeFeatureGenerator())
-
-    # Individual generators
-    timeframes = [1, 5, 15, 30, 60]
-    feature_types = ["volatility", "momentum", "volume", "trend"]
-
-    # Fractional change generators
-    for fast_tf in timeframes:
-        for slow_tf in timeframes:
-            if fast_tf < slow_tf:
-                for feature_type in feature_types:
-                    generators.append(FractionalChangeGenerator(fast_tf=fast_tf, slow_tf=slow_tf, feature_type=feature_type))
-
-    # Alignment generators
-    for source_tf in [1, 5]:
-        for target_tf in [5, 15, 30]:
-            if source_tf < target_tf:
-                for method in ["lag", "resample"]:
-                    generators.append(CrossTimeframeAlignmentGenerator(source_tf=source_tf, target_tf=target_tf, alignment_method=method))
-
-    # Learned projection generators
-    for n_components in [2, 3, 5]:
-        generators.append(LearnedProjectionGenerator(timeframes=[1, 5, 15], n_components=n_components))
-
-    return generators
-
-
-def create_default_enhanced_cross_timeframe_generators() -> List[FeatureGenerator]:
-    """Create default set of enhanced cross-timeframe generators."""
-    return create_enhanced_cross_timeframe_generators()
-
-
-# Export all generators
-__all__ = [
-    'EnhancedCrossTimeframeFeatureGenerator',
-    'FractionalChangeGenerator',
-    'CrossTimeframeAlignmentGenerator',
-    'LearnedProjectionGenerator',
-    'create_enhanced_cross_timeframe_generators',
-    'create_default_enhanced_cross_timeframe_generators'
-]
