--- src/feature_generation/categories/enhanced_representation_learning.py
@@ -1,1014 +1,0 @@
-"""
-Enhanced Representation Learning Feature Generator
-
-This module provides advanced representation learning features using
-PatchTST, TFT encoders, autoencoders, and contrastive learning to
-create learned embeddings that summarize market dynamics for use in
-tree-based models.
-
-Features implemented:
-- PatchTST self-supervised learning with proper masking
-- TFT encoder representations with attention mechanisms
-- Autoencoder-based embeddings with multiple architectures
-- Contrastive learning representations
-- Multi-scale representation learning
-- Regime-aware representation learning
-- Cross-timeframe representation fusion
-"""
-
-import numpy as np
-import pandas as pd
-from typing import Any, Dict, List, Optional, Union, Tuple
-import logging
-from sklearn.decomposition import PCA, FastICA
-from sklearn.manifold import TSNE
-from sklearn.preprocessing import StandardScaler
-import warnings
-warnings.filterwarnings('ignore')
-
-from ..core.feature_generator import (
-    FeatureGenerator,
-    FeatureConfig,
-    FeatureCategory,
-    VectorizedFeatureGenerator
-)
-
-logger = logging.getLogger(__name__)
-
-
-class EnhancedRepresentationLearningGenerator(VectorizedFeatureGenerator):
-    """Enhanced feature generator for representation learning features."""
-
-    def __init__(self, config: Optional[FeatureConfig] = None):
-        if config is None:
-            config = self._create_default_config()
-        super().__init__(config, enable_matrix_ops=True)
-
-    @classmethod
-    def _create_default_config(cls) -> FeatureConfig:
-        return FeatureConfig(
-            name="enhanced_representation_learning_features",
-            category=FeatureCategory.REPRESENTATION_LEARNING,
-            description="Enhanced representation learning features with multiple architectures",
-            required_columns=["close"],
-            optional_columns=["high", "low", "open", "volume"],
-            default_lookback=200,
-            min_lookback=100,
-            max_lookback=1000,
-            parameters={
-                "patch_lengths": [8, 16, 32],
-                "num_patches": [4, 8, 16],
-                "embedding_dims": [32, 64, 128],
-                "sequence_lengths": [60, 120, 240],
-                "masking_ratios": [0.15, 0.3, 0.5],
-                "representation_methods": ["patchtst", "tft", "autoencoder", "contrastive", "pca", "ica"],
-                "regime_aware": True,
-                "multi_scale": True,
-                "cross_timeframe": True
-            },
-            matrix_optimized=True,
-            gpu_accelerated=False
-        )
-
-    def _generate_feature(self, data: pd.DataFrame, **kwargs) -> pd.Series:
-        """Generate enhanced representation learning features."""
-        try:
-            # Generate all enhanced representation learning features
-            features_dict = self.generate_enhanced_representation_features(data, **kwargs)
-
-            # Return first feature as representative for base class
-            if features_dict:
-                first_feature_name = list(features_dict.keys())[0]
-                return pd.Series(features_dict[first_feature_name], index=data.index)
-            else:
-                return pd.Series(np.zeros(len(data)), index=data.index)
-
-        except Exception as e:
-            logger.error(f"Error generating enhanced representation learning features: {e}")
-            return pd.Series(np.zeros(len(data)), index=data.index)
-
-    def generate_enhanced_representation_features(self, data: pd.DataFrame, **kwargs) -> Dict[str, np.ndarray]:
-        """Generate comprehensive enhanced representation learning features."""
-        features = {}
-
-        try:
-            # PatchTST representation features
-            features.update(self._generate_patchtst_features(data))
-
-            # TFT encoder representation features
-            features.update(self._generate_tft_features(data))
-
-            # Autoencoder representation features
-            features.update(self._generate_autoencoder_features(data))
-
-            # Contrastive learning features
-            features.update(self._generate_contrastive_features(data))
-
-            # PCA/ICA representation features
-            features.update(self._generate_dimensionality_reduction_features(data))
-
-            # Multi-scale representation features
-            features.update(self._generate_multiscale_representation_features(data))
-
-            # Regime-aware representation features
-            features.update(self._generate_regime_aware_representation_features(data))
-
-            # Cross-timeframe representation features
-            features.update(self._generate_cross_timeframe_representation_features(data))
-
-            logger.info(f"Generated {len(features)} enhanced representation learning features")
-            return features
-
-        except Exception as e:
-            logger.error(f"Error in generate_enhanced_representation_features: {e}")
-            return {}
-
-    def _generate_patchtst_features(self, data: pd.DataFrame) -> Dict[str, np.ndarray]:
-        """Generate PatchTST representation features."""
-        features = {}
-
-        patch_lengths = self.config.parameters.get("patch_lengths", [8, 16, 32])
-        num_patches = self.config.parameters.get("num_patches", [4, 8, 16])
-        embedding_dims = self.config.parameters.get("embedding_dims", [32, 64, 128])
-        masking_ratios = self.config.parameters.get("masking_ratios", [0.15, 0.3, 0.5])
-
-        for patch_length in patch_lengths:
-            for num_patch in num_patches:
-                for embedding_dim in embedding_dims:
-                    for masking_ratio in masking_ratios:
-                        try:
-                            # Create patches
-                            patches = self._create_patchtst_patches(data["close"], patch_length, num_patch)
-                            
-                            if patches is not None:
-                                # Apply masking
-                                masked_patches, mask = self._apply_patchtst_masking(patches, masking_ratio)
-                                
-                                # Learn representations
-                                representations = self._learn_patchtst_representations(masked_patches, embedding_dim)
-                                
-                                # Store features
-                                for i in range(min(embedding_dim, representations.shape[1])):
-                                    features[f"patchtst_{patch_length}_{num_patch}_{embedding_dim}_{masking_ratio}_comp_{i}"] = representations[:, i]
-
-                        except Exception as e:
-                            logger.warning(f"Error in PatchTST {patch_length}_{num_patch}_{embedding_dim}: {e}")
-
-        return features
-
-    def _create_patchtst_patches(self, series: pd.Series, patch_length: int, num_patches: int) -> Optional[np.ndarray]:
-        """Create patches for PatchTST."""
-        seq_len = len(series)
-        patch_size = patch_length * num_patches
-
-        if seq_len < patch_size:
-            return None
-
-        # Take the most recent data
-        recent_data = series.values[-patch_size:]
-        
-        # Reshape into patches
-        patches = recent_data.reshape(num_patches, patch_length)
-        return patches
-
-    def _apply_patchtst_masking(self, patches: np.ndarray, masking_ratio: float) -> Tuple[np.ndarray, np.ndarray]:
-        """Apply masking for PatchTST self-supervised learning."""
-        masked_patches = patches.copy()
-        mask = np.zeros_like(patches, dtype=bool)
-
-        # Random masking
-        total_elements = patches.size
-        num_masked = int(total_elements * masking_ratio)
-
-        if num_masked > 0:
-            masked_indices = np.random.choice(total_elements, num_masked, replace=False)
-            mask.flat[masked_indices] = True
-            masked_patches.flat[masked_indices] = 0.0
-
-        return masked_patches, mask
-
-    def _learn_patchtst_representations(self, patches: np.ndarray, embedding_dim: int) -> np.ndarray:
-        """Learn patch representations using statistical features."""
-        # Calculate statistical features for each patch
-        patch_means = patches.mean(axis=1)
-        patch_stds = patches.std(axis=1)
-        patch_trends = np.polyfit(np.arange(patches.shape[1]), patches.T, 1)[0]
-        patch_curvatures = np.polyfit(np.arange(patches.shape[1]), patches.T, 2)[0]
-
-        # Combine into representation vectors
-        base_features = np.column_stack([patch_means, patch_stds, patch_trends, patch_curvatures])
-        
-        # Pad or truncate to embedding_dim
-        if base_features.shape[1] < embedding_dim:
-            # Pad with zeros
-            padding = np.zeros((base_features.shape[0], embedding_dim - base_features.shape[1]))
-            representations = np.column_stack([base_features, padding])
-        else:
-            # Truncate
-            representations = base_features[:, :embedding_dim]
-
-        return representations
-
-    def _generate_tft_features(self, data: pd.DataFrame) -> Dict[str, np.ndarray]:
-        """Generate TFT encoder representation features."""
-        features = {}
-
-        sequence_lengths = self.config.parameters.get("sequence_lengths", [60, 120, 240])
-        embedding_dims = self.config.parameters.get("embedding_dims", [32, 64, 128])
-
-        for seq_length in sequence_lengths:
-            for embedding_dim in embedding_dims:
-                try:
-                    # Extract multi-variate time series
-                    input_features = self._extract_tft_input_features(data, seq_length)
-                    
-                    if input_features is not None:
-                        # Apply self-attention mechanism
-                        attention_output = self._apply_tft_attention(input_features)
-                        
-                        # Generate temporal representations
-                        temporal_repr = self._tft_temporal_fusion(attention_output, embedding_dim)
-                        
-                        # Store features
-                        for i in range(min(embedding_dim, temporal_repr.shape[1])):
-                            features[f"tft_{seq_length}_{embedding_dim}_comp_{i}"] = temporal_repr[:, i]
-
-                except Exception as e:
-                    logger.warning(f"Error in TFT {seq_length}_{embedding_dim}: {e}")
-
-        return features
-
-    def _extract_tft_input_features(self, data: pd.DataFrame, seq_length: int) -> Optional[np.ndarray]:
-        """Extract input features for TFT."""
-        try:
-            features = []
-
-            # Price features
-            price_features = [
-                data["close"].pct_change().fillna(0),
-                (data["close"] - data["close"].rolling(window=20).mean()) / (data["close"].rolling(window=20).std() + 1e-8),
-                data["close"].rolling(window=10).mean() / data["close"].rolling(window=30).mean() - 1,
-            ]
-
-            # Volatility features
-            returns = data["close"].pct_change()
-            volatility_features = [
-                returns.rolling(window=20).std().fillna(0),
-                returns.rolling(window=5).std() / (returns.rolling(window=20).std() + 1e-8),
-            ]
-
-            # Volume features (if available)
-            if "volume" in data.columns:
-                volume_features = [
-                    data["volume"] / (data["volume"].rolling(window=20).mean() + 1e-8),
-                    data["volume"].pct_change().fillna(0),
-                ]
-            else:
-                volume_features = [np.zeros(len(data)), np.zeros(len(data))]
-
-            # Combine all features
-            all_features = price_features + volatility_features + volume_features
-            feature_matrix = np.column_stack([f.fillna(0).values for f in all_features])
-
-            # Truncate to sequence length
-            if len(feature_matrix) > seq_length:
-                feature_matrix = feature_matrix[-seq_length:]
-
-            return feature_matrix
-
-        except Exception as e:
-            logger.warning(f"Error extracting TFT input features: {e}")
-            return None
-
-    def _apply_tft_attention(self, features: np.ndarray) -> np.ndarray:
-        """Apply self-attention mechanism for TFT."""
-        try:
-            # Simple attention mechanism
-            query = features
-            key = features
-            value = features
-
-            # Attention scores (using cosine similarity as proxy)
-            attention_scores = np.dot(query, key.T) / (np.linalg.norm(query, axis=1, keepdims=True) * np.linalg.norm(key, axis=1, keepdims=True).T + 1e-8)
-
-            # Apply softmax
-            attention_weights = np.exp(attention_scores) / (np.sum(np.exp(attention_scores), axis=1, keepdims=True) + 1e-8)
-
-            # Weighted sum
-            attention_output = np.dot(attention_weights, value)
-
-            return attention_output
-
-        except Exception as e:
-            logger.warning(f"Error in TFT attention: {e}")
-            return features
-
-    def _tft_temporal_fusion(self, attention_output: np.ndarray, embedding_dim: int) -> np.ndarray:
-        """Apply temporal fusion to generate final representations."""
-        try:
-            # Use rolling statistics as temporal representation
-            window_sizes = [5, 10, 20]
-            temporal_features = []
-
-            for window in window_sizes:
-                if len(attention_output) >= window:
-                    # Rolling mean
-                    rolling_mean = pd.Series(attention_output.mean(axis=1)).rolling(window=window).mean().fillna(0)
-                    temporal_features.append(rolling_mean.values)
-
-                    # Rolling std
-                    rolling_std = pd.Series(attention_output.mean(axis=1)).rolling(window=window).std().fillna(0)
-                    temporal_features.append(rolling_std.values)
-
-            if temporal_features:
-                combined_features = np.column_stack(temporal_features)
-                
-                # Pad or truncate to embedding_dim
-                if combined_features.shape[1] < embedding_dim:
-                    padding = np.zeros((combined_features.shape[0], embedding_dim - combined_features.shape[1]))
-                    return np.column_stack([combined_features, padding])
-                else:
-                    return combined_features[:, :embedding_dim]
-            else:
-                return attention_output
-
-        except Exception as e:
-            logger.warning(f"Error in TFT temporal fusion: {e}")
-            return attention_output
-
-    def _generate_autoencoder_features(self, data: pd.DataFrame) -> Dict[str, np.ndarray]:
-        """Generate autoencoder representation features."""
-        features = {}
-
-        sequence_lengths = self.config.parameters.get("sequence_lengths", [60, 120, 240])
-        embedding_dims = self.config.parameters.get("embedding_dims", [32, 64, 128])
-
-        for seq_length in sequence_lengths:
-            for embedding_dim in embedding_dims:
-                try:
-                    # Create input features
-                    input_features = self._create_autoencoder_input_features(data, seq_length)
-                    
-                    if input_features is not None:
-                        # Apply autoencoder (using PCA as proxy)
-                        encoded_repr = self._encode_autoencoder_features(input_features, embedding_dim)
-                        
-                        # Store features
-                        for i in range(min(embedding_dim, encoded_repr.shape[1])):
-                            features[f"autoencoder_{seq_length}_{embedding_dim}_comp_{i}"] = encoded_repr[:, i]
-
-                except Exception as e:
-                    logger.warning(f"Error in autoencoder {seq_length}_{embedding_dim}: {e}")
-
-        return features
-
-    def _create_autoencoder_input_features(self, data: pd.DataFrame, seq_length: int) -> Optional[np.ndarray]:
-        """Create input features for autoencoder."""
-        try:
-            features = []
-
-            # Technical indicators as features
-            indicators = [
-                data["close"].pct_change().fillna(0),
-                data["close"].rolling(window=10).mean().fillna(data["close"]),
-                data["close"].rolling(window=20).std().fillna(0),
-                data["close"] / data["close"].rolling(window=10).mean() - 1,
-            ]
-
-            if "volume" in data.columns:
-                indicators.extend([
-                    data["volume"] / (data["volume"].rolling(window=20).mean() + 1e-8),
-                    data["volume"].pct_change().fillna(0),
-                ])
-
-            # Stack features
-            feature_matrix = np.column_stack([ind.fillna(0).values for ind in indicators])
-
-            # Ensure we have enough data
-            if feature_matrix.shape[0] > seq_length:
-                feature_matrix = feature_matrix[-seq_length:]
-
-            return feature_matrix
-
-        except Exception as e:
-            logger.warning(f"Error creating autoencoder input features: {e}")
-            return None
-
-    def _encode_autoencoder_features(self, features: np.ndarray, embedding_dim: int) -> np.ndarray:
-        """Encode features using autoencoder (PCA as proxy)."""
-        try:
-            # Reshape for time series encoding
-            seq_len, feature_dim = features.shape
-
-            # Flatten for PCA
-            flattened = features.flatten().reshape(1, -1)
-
-            # Apply PCA for dimensionality reduction
-            if feature_dim >= embedding_dim:
-                pca = PCA(n_components=embedding_dim)
-                encoded = pca.fit_transform(flattened)
-            else:
-                # Pad if needed
-                encoded = np.pad(flattened, ((0, 0), (0, embedding_dim - feature_dim)), mode='constant')
-
-            # Reshape back to sequence format
-            encoded_reshaped = encoded.reshape(seq_len, -1)
-
-            return encoded_reshaped
-
-        except Exception as e:
-            logger.warning(f"Error in autoencoder encoding: {e}")
-            return features.mean(axis=1, keepdims=True)
-
-    def _generate_contrastive_features(self, data: pd.DataFrame) -> Dict[str, np.ndarray]:
-        """Generate contrastive learning representation features."""
-        features = {}
-
-        embedding_dims = self.config.parameters.get("embedding_dims", [32, 64, 128])
-        temperatures = [0.1, 0.5, 1.0]
-
-        for embedding_dim in embedding_dims:
-            for temperature in temperatures:
-                try:
-                    # Create positive and negative samples
-                    positive_samples, negative_samples = self._create_contrastive_samples(data)
-                    
-                    if len(positive_samples) > 0:
-                        # Learn representations through contrastive loss
-                        representations = self._contrastive_learning(positive_samples, negative_samples, embedding_dim, temperature)
-                        
-                        # Store features
-                        for i in range(min(embedding_dim, representations.shape[1])):
-                            features[f"contrastive_{embedding_dim}_{temperature}_comp_{i}"] = representations[:, i]
-
-                except Exception as e:
-                    logger.warning(f"Error in contrastive learning {embedding_dim}_{temperature}: {e}")
-
-        return features
-
-    def _create_contrastive_samples(self, data: pd.DataFrame) -> Tuple[List[Tuple[np.ndarray, np.ndarray]], List[Tuple[np.ndarray, np.ndarray]]]:
-        """Create positive and negative samples for contrastive learning."""
-        try:
-            # Extract price sequence
-            price_sequence = data["close"].values
-
-            # Create overlapping windows as positive pairs
-            window_size = 20
-            stride = 5
-
-            positive_samples = []
-            for i in range(0, len(price_sequence) - window_size * 2, stride):
-                # Positive pair: consecutive windows
-                window1 = price_sequence[i:i+window_size]
-                window2 = price_sequence[i+stride:i+stride+window_size]
-
-                if len(window1) == window_size and len(window2) == window_size:
-                    positive_samples.append((window1, window2))
-
-            # Negative samples: random windows
-            negative_samples = []
-            num_negative = min(len(positive_samples), 100)  # Limit to prevent memory issues
-
-            for _ in range(num_negative):
-                # Randomly sample two different windows
-                if len(price_sequence) > window_size * 2:
-                    idx1, idx2 = np.random.choice(len(price_sequence) - window_size, 2, replace=False)
-                    neg1 = price_sequence[idx1:idx1+window_size]
-                    neg2 = price_sequence[idx2:idx2+window_size]
-                    negative_samples.append((neg1, neg2))
-
-            return positive_samples, negative_samples
-
-        except Exception as e:
-            logger.warning(f"Error creating contrastive samples: {e}")
-            return [], []
-
-    def _contrastive_learning(self, positive_samples: List[Tuple[np.ndarray, np.ndarray]], 
-                            negative_samples: List[Tuple[np.ndarray, np.ndarray]], 
-                            embedding_dim: int, temperature: float) -> np.ndarray:
-        """Apply contrastive learning."""
-        try:
-            representations = []
-
-            for pos_pair, neg_pair in zip(positive_samples, negative_samples):
-                # Calculate similarities
-                pos_sim = np.corrcoef(pos_pair[0], pos_pair[1])[0, 1] if len(pos_pair[0]) > 1 and len(pos_pair[1]) > 1 else 0
-                neg_sim = np.corrcoef(neg_pair[0], neg_pair[1])[0, 1] if len(neg_pair[0]) > 1 and len(neg_pair[1]) > 1 else 0
-
-                # Contrastive representation
-                contrastive_repr = pos_sim - neg_sim
-                representations.append(contrastive_repr)
-
-            if representations:
-                # Convert to array and pad/truncate to embedding_dim
-                repr_array = np.array(representations)
-                if len(repr_array) < embedding_dim:
-                    padding = np.zeros(embedding_dim - len(repr_array))
-                    return np.concatenate([repr_array, padding]).reshape(1, -1)
-                else:
-                    return repr_array[:embedding_dim].reshape(1, -1)
-            else:
-                return np.zeros((1, embedding_dim))
-
-        except Exception as e:
-            logger.warning(f"Error in contrastive learning: {e}")
-            return np.zeros((1, embedding_dim))
-
-    def _generate_dimensionality_reduction_features(self, data: pd.DataFrame) -> Dict[str, np.ndarray]:
-        """Generate PCA/ICA representation features."""
-        features = {}
-
-        embedding_dims = self.config.parameters.get("embedding_dims", [32, 64, 128])
-
-        for embedding_dim in embedding_dims:
-            try:
-                # Create feature matrix
-                feature_matrix = self._create_dimensionality_reduction_input(data)
-                
-                if feature_matrix is not None and feature_matrix.shape[1] >= embedding_dim:
-                    # PCA
-                    pca = PCA(n_components=embedding_dim)
-                    pca_result = pca.fit_transform(feature_matrix)
-                    
-                    for i in range(embedding_dim):
-                        features[f"pca_{embedding_dim}_comp_{i}"] = pca_result[:, i]
-
-                    # ICA
-                    ica = FastICA(n_components=embedding_dim, random_state=42)
-                    ica_result = ica.fit_transform(feature_matrix)
-                    
-                    for i in range(embedding_dim):
-                        features[f"ica_{embedding_dim}_comp_{i}"] = ica_result[:, i]
-
-            except Exception as e:
-                logger.warning(f"Error in dimensionality reduction {embedding_dim}: {e}")
-
-        return features
-
-    def _create_dimensionality_reduction_input(self, data: pd.DataFrame) -> Optional[np.ndarray]:
-        """Create input features for dimensionality reduction."""
-        try:
-            features = []
-
-            # Price features
-            features.extend([
-                data["close"].pct_change().fillna(0),
-                data["close"].rolling(window=5).mean().fillna(data["close"]),
-                data["close"].rolling(window=10).mean().fillna(data["close"]),
-                data["close"].rolling(window=20).mean().fillna(data["close"]),
-            ])
-
-            # Volatility features
-            returns = data["close"].pct_change()
-            features.extend([
-                returns.rolling(window=5).std().fillna(0),
-                returns.rolling(window=10).std().fillna(0),
-                returns.rolling(window=20).std().fillna(0),
-            ])
-
-            # Volume features
-            if "volume" in data.columns:
-                features.extend([
-                    data["volume"].pct_change().fillna(0),
-                    data["volume"] / (data["volume"].rolling(window=20).mean() + 1e-8),
-                ])
-
-            # High-low features
-            if "high" in data.columns and "low" in data.columns:
-                features.extend([
-                    (data["high"] - data["low"]) / data["close"],
-                    (data["close"] - data["low"]) / (data["high"] - data["low"] + 1e-8),
-                ])
-
-            # Combine features
-            feature_matrix = np.column_stack([f.fillna(0).values for f in features])
-            
-            return feature_matrix
-
-        except Exception as e:
-            logger.warning(f"Error creating dimensionality reduction input: {e}")
-            return None
-
-    def _generate_multiscale_representation_features(self, data: pd.DataFrame) -> Dict[str, np.ndarray]:
-        """Generate multi-scale representation features."""
-        features = {}
-
-        scales = [5, 10, 20, 50]
-        embedding_dims = [16, 32, 64]
-
-        for scale in scales:
-            for embedding_dim in embedding_dims:
-                try:
-                    # Create multi-scale features
-                    scale_features = self._create_multiscale_features(data, scale)
-                    
-                    if scale_features is not None:
-                        # Apply PCA for dimensionality reduction
-                        pca = PCA(n_components=min(embedding_dim, scale_features.shape[1]))
-                        pca_result = pca.fit_transform(scale_features)
-                        
-                        for i in range(pca_result.shape[1]):
-                            features[f"multiscale_{scale}_{embedding_dim}_comp_{i}"] = pca_result[:, i]
-
-                except Exception as e:
-                    logger.warning(f"Error in multiscale {scale}_{embedding_dim}: {e}")
-
-        return features
-
-    def _create_multiscale_features(self, data: pd.DataFrame, scale: int) -> Optional[np.ndarray]:
-        """Create multi-scale features."""
-        try:
-            features = []
-
-            # Different scales of the same feature
-            for window in [scale//2, scale, scale*2]:
-                if window > 0:
-                    features.extend([
-                        data["close"].pct_change(window).fillna(0),
-                        data["close"].rolling(window=window).mean().fillna(data["close"]),
-                        data["close"].pct_change().rolling(window=window).std().fillna(0),
-                    ])
-
-            if features:
-                return np.column_stack([f.fillna(0).values for f in features])
-            else:
-                return None
-
-        except Exception as e:
-            logger.warning(f"Error creating multiscale features: {e}")
-            return None
-
-    def _generate_regime_aware_representation_features(self, data: pd.DataFrame) -> Dict[str, np.ndarray]:
-        """Generate regime-aware representation features."""
-        features = {}
-
-        # Detect regimes
-        regimes = self._detect_regimes(data)
-
-        for regime_name, regime_mask in regimes.items():
-            if regime_mask.sum() > 10:  # Need sufficient data
-                try:
-                    # Create regime-specific features
-                    regime_data = data[regime_mask]
-                    regime_features = self._create_dimensionality_reduction_input(regime_data)
-                    
-                    if regime_features is not None and regime_features.shape[0] > 5:
-                        # Apply PCA
-                        pca = PCA(n_components=min(3, regime_features.shape[1]))
-                        pca_result = pca.fit_transform(regime_features)
-                        
-                        # Create full-length feature with regime-specific values
-                        full_feature = np.zeros(len(data))
-                        full_feature[regime_mask] = pca_result[:, 0] if pca_result.shape[1] > 0 else 0
-                        
-                        features[f"regime_{regime_name}_repr"] = full_feature
-
-                except Exception as e:
-                    logger.warning(f"Error in regime-aware representation {regime_name}: {e}")
-
-        return features
-
-    def _detect_regimes(self, data: pd.DataFrame) -> Dict[str, pd.Series]:
-        """Detect market regimes."""
-        regimes = {}
-
-        # Volatility regime
-        returns = data["close"].pct_change()
-        vol = returns.rolling(window=20).std()
-        vol_percentiles = vol.quantile([0.33, 0.67])
-        
-        regimes["low_vol"] = (vol <= vol_percentiles.iloc[0]).astype(int)
-        regimes["high_vol"] = (vol >= vol_percentiles.iloc[1]).astype(int)
-        regimes["normal_vol"] = ((vol > vol_percentiles.iloc[0]) & (vol < vol_percentiles.iloc[1])).astype(int)
-
-        # Momentum regime
-        momentum = data["close"].pct_change(20)
-        mom_percentiles = momentum.quantile([0.33, 0.67])
-        
-        regimes["uptrend"] = (momentum >= mom_percentiles.iloc[1]).astype(int)
-        regimes["downtrend"] = (momentum <= mom_percentiles.iloc[0]).astype(int)
-
-        return regimes
-
-    def _generate_cross_timeframe_representation_features(self, data: pd.DataFrame) -> Dict[str, np.ndarray]:
-        """Generate cross-timeframe representation features."""
-        features = {}
-
-        timeframes = [5, 15, 30, 60]
-        embedding_dims = [16, 32]
-
-        for embedding_dim in embedding_dims:
-            try:
-                # Create cross-timeframe features
-                tf_features = []
-                for tf in timeframes:
-                    tf_returns = data["close"].pct_change(tf).fillna(0)
-                    tf_vol = data["close"].pct_change().rolling(window=tf).std().fillna(0)
-                    tf_features.extend([tf_returns, tf_vol])
-
-                if tf_features:
-                    feature_matrix = np.column_stack([f.fillna(0).values for f in tf_features])
-                    
-                    # Apply PCA
-                    pca = PCA(n_components=min(embedding_dim, feature_matrix.shape[1]))
-                    pca_result = pca.fit_transform(feature_matrix)
-                    
-                    for i in range(pca_result.shape[1]):
-                        features[f"cross_tf_{embedding_dim}_comp_{i}"] = pca_result[:, i]
-
-            except Exception as e:
-                logger.warning(f"Error in cross-timeframe representation {embedding_dim}: {e}")
-
-        return features
-
-
-# Individual enhanced representation learning generators
-
-class PatchTSTRepresentationGenerator(FeatureGenerator):
-    """Generator for PatchTST-based representation learning."""
-
-    def __init__(self, patch_length: int = 16, num_patches: int = 8, embedding_dim: int = 64):
-        config = FeatureConfig(
-            name=f"patchtst_repr_{patch_length}_{num_patches}_{embedding_dim}",
-            category=FeatureCategory.REPRESENTATION_LEARNING,
-            description=f"PatchTST representation learning with patch_length={patch_length}, num_patches={num_patches}, embedding_dim={embedding_dim}",
-            required_columns=["close"],
-            optional_columns=["high", "low", "open", "volume"],
-            default_lookback=patch_length * num_patches * 2,
-            min_lookback=patch_length * num_patches,
-            max_lookback=patch_length * num_patches * 4,
-            parameters={
-                "patch_length": patch_length,
-                "num_patches": num_patches,
-                "embedding_dim": embedding_dim,
-                "masking_ratio": 0.5
-            }
-        )
-        super().__init__(config)
-        self.patch_length = patch_length
-        self.num_patches = num_patches
-        self.embedding_dim = embedding_dim
-        self.masking_ratio = 0.5
-
-    def _generate_feature(self, data: pd.DataFrame, **kwargs) -> pd.Series:
-        """Generate PatchTST representation feature."""
-        try:
-            # Extract price sequence
-            price_sequence = data["close"].values
-
-            # Create patches
-            patches = self._create_patches(price_sequence)
-            
-            if patches is None:
-                return pd.Series(np.zeros(len(data)), index=data.index)
-
-            # Apply masking for self-supervised learning
-            masked_patches, mask = self._apply_masking(patches)
-
-            # Learn representations
-            representations = self._learn_patch_representations(masked_patches)
-
-            # Return concatenated representation as single feature
-            return pd.Series(representations.mean(axis=1), index=data.index[-len(representations):])
-
-        except Exception as e:
-            logger.warning(f"Error in PatchTST representation generation: {e}")
-            return pd.Series(np.zeros(len(data)), index=data.index)
-
-    def _create_patches(self, sequence: np.ndarray) -> Optional[np.ndarray]:
-        """Create patches from price sequence."""
-        seq_len = len(sequence)
-        patch_size = self.patch_length * self.num_patches
-
-        if seq_len < patch_size:
-            return None
-
-        # Take the most recent data
-        recent_data = sequence[-patch_size:]
-        
-        # Reshape into patches
-        patches = recent_data.reshape(self.num_patches, self.patch_length)
-        return patches
-
-    def _apply_masking(self, patches: np.ndarray) -> Tuple[np.ndarray, np.ndarray]:
-        """Apply masking for self-supervised learning."""
-        masked_patches = patches.copy()
-        mask = np.zeros_like(patches, dtype=bool)
-
-        # Random masking
-        total_elements = patches.size
-        num_masked = int(total_elements * self.masking_ratio)
-
-        if num_masked > 0:
-            masked_indices = np.random.choice(total_elements, num_masked, replace=False)
-            mask.flat[masked_indices] = True
-            masked_patches.flat[masked_indices] = 0.0
-
-        return masked_patches, mask
-
-    def _learn_patch_representations(self, patches: np.ndarray) -> np.ndarray:
-        """Learn patch representations using statistical features."""
-        # Calculate statistical features for each patch
-        patch_means = patches.mean(axis=1)
-        patch_stds = patches.std(axis=1)
-        patch_trends = np.polyfit(np.arange(self.patch_length), patches.T, 1)[0]
-        patch_curvatures = np.polyfit(np.arange(self.patch_length), patches.T, 2)[0]
-
-        # Combine into representation vectors
-        representations = np.column_stack([patch_means, patch_stds, patch_trends, patch_curvatures])
-        
-        # Pad or truncate to embedding_dim
-        if representations.shape[1] < self.embedding_dim:
-            padding = np.zeros((representations.shape[0], self.embedding_dim - representations.shape[1]))
-            representations = np.column_stack([representations, padding])
-        else:
-            representations = representations[:, :self.embedding_dim]
-
-        return representations
-
-
-class TFTEncoderRepresentationGenerator(FeatureGenerator):
-    """Generator for TFT encoder representation learning."""
-
-    def __init__(self, seq_length: int = 60, hidden_size: int = 64, num_heads: int = 4):
-        config = FeatureConfig(
-            name=f"tft_encoder_repr_{seq_length}_{hidden_size}_{num_heads}",
-            category=FeatureCategory.REPRESENTATION_LEARNING,
-            description=f"TFT encoder representation learning with seq_length={seq_length}, hidden_size={hidden_size}",
-            required_columns=["close"],
-            optional_columns=["high", "low", "open", "volume"],
-            default_lookback=seq_length * 2,
-            min_lookback=seq_length,
-            max_lookback=seq_length * 4,
-            parameters={
-                "seq_length": seq_length,
-                "hidden_size": hidden_size,
-                "num_heads": num_heads
-            }
-        )
-        super().__init__(config)
-        self.seq_length = seq_length
-        self.hidden_size = hidden_size
-        self.num_heads = num_heads
-
-    def _generate_feature(self, data: pd.DataFrame, **kwargs) -> pd.Series:
-        """Generate TFT encoder representation feature."""
-        try:
-            # Extract multi-variate time series
-            features = self._extract_features(data)
-
-            if features is None:
-                return pd.Series(np.zeros(len(data)), index=data.index)
-
-            # Apply self-attention mechanism
-            attention_output = self._apply_self_attention(features)
-
-            # Generate temporal representations
-            temporal_repr = self._temporal_fusion(attention_output)
-
-            # Return representation as feature
-            return pd.Series(temporal_repr.mean(axis=1), index=data.index[-len(temporal_repr):])
-
-        except Exception as e:
-            logger.warning(f"Error in TFT encoder representation generation: {e}")
-            return pd.Series(np.zeros(len(data)), index=data.index)
-
-    def _extract_features(self, data: pd.DataFrame) -> Optional[np.ndarray]:
-        """Extract features for TFT input."""
-        try:
-            features = []
-
-            # Price features
-            price_features = [
-                data["close"].pct_change().fillna(0),
-                (data["close"] - data["close"].rolling(window=20).mean()) / (data["close"].rolling(window=20).std() + 1e-8),
-                data["close"].rolling(window=10).mean() / data["close"].rolling(window=30).mean() - 1,
-            ]
-
-            # Volatility features
-            returns = data["close"].pct_change()
-            volatility_features = [
-                returns.rolling(window=20).std().fillna(0),
-                returns.rolling(window=5).std() / (returns.rolling(window=20).std() + 1e-8),
-            ]
-
-            # Volume features (if available)
-            if "volume" in data.columns:
-                volume_features = [
-                    data["volume"] / (data["volume"].rolling(window=20).mean() + 1e-8),
-                    data["volume"].pct_change().fillna(0),
-                ]
-            else:
-                volume_features = [np.zeros(len(data)), np.zeros(len(data))]
-
-            # Combine all features
-            all_features = price_features + volatility_features + volume_features
-            feature_matrix = np.column_stack([f.fillna(0).values for f in all_features])
-
-            # Truncate to sequence length
-            if len(feature_matrix) > self.seq_length:
-                feature_matrix = feature_matrix[-self.seq_length:]
-
-            return feature_matrix
-
-        except Exception as e:
-            logger.warning(f"Error extracting TFT features: {e}")
-            return None
-
-    def _apply_self_attention(self, features: np.ndarray) -> np.ndarray:
-        """Apply self-attention mechanism."""
-        try:
-            # Simple attention mechanism
-            query = features
-            key = features
-            value = features
-
-            # Attention scores (using cosine similarity as proxy)
-            attention_scores = np.dot(query, key.T) / (np.linalg.norm(query, axis=1, keepdims=True) * np.linalg.norm(key, axis=1, keepdims=True).T + 1e-8)
-
-            # Apply softmax
-            attention_weights = np.exp(attention_scores) / (np.sum(np.exp(attention_scores), axis=1, keepdims=True) + 1e-8)
-
-            # Weighted sum
-            attention_output = np.dot(attention_weights, value)
-
-            return attention_output
-
-        except Exception as e:
-            logger.warning(f"Error in self-attention: {e}")
-            return features
-
-    def _temporal_fusion(self, attention_output: np.ndarray) -> np.ndarray:
-        """Apply temporal fusion to generate final representations."""
-        try:
-            # Use rolling statistics as temporal representation
-            window_sizes = [5, 10, 20]
-            temporal_features = []
-
-            for window in window_sizes:
-                if len(attention_output) >= window:
-                    # Rolling mean
-                    rolling_mean = pd.Series(attention_output.mean(axis=1)).rolling(window=window).mean().fillna(0)
-                    temporal_features.append(rolling_mean.values)
-
-                    # Rolling std
-                    rolling_std = pd.Series(attention_output.mean(axis=1)).rolling(window=window).std().fillna(0)
-                    temporal_features.append(rolling_std.values)
-
-            if temporal_features:
-                return np.column_stack(temporal_features)
-            else:
-                return attention_output
-
-        except Exception as e:
-            logger.warning(f"Error in temporal fusion: {e}")
-            return attention_output
-
-
-def create_enhanced_representation_learning_generators() -> List[FeatureGenerator]:
-    """Create all enhanced representation learning feature generators."""
-    generators = []
-
-    # Main enhanced representation learning generator
-    generators.append(EnhancedRepresentationLearningGenerator())
-
-    # Individual generators
-    patch_configs = [
-        (8, 4, 32), (16, 8, 64), (32, 16, 128)
-    ]
-
-    for patch_length, num_patches, embedding_dim in patch_configs:
-        generators.append(PatchTSTRepresentationGenerator(
-            patch_length=patch_length,
-            num_patches=num_patches,
-            embedding_dim=embedding_dim
-        ))
-
-    # TFT generators
-    tft_configs = [
-        (60, 64, 4), (120, 128, 8), (240, 256, 16)
-    ]
-
-    for seq_length, hidden_size, num_heads in tft_configs:
-        generators.append(TFTEncoderRepresentationGenerator(
-            seq_length=seq_length,
-            hidden_size=hidden_size,
-            num_heads=num_heads
-        ))
-
-    return generators
-
-
-def create_default_enhanced_representation_learning_generators() -> List[FeatureGenerator]:
-    """Create default set of enhanced representation learning generators."""
-    return create_enhanced_representation_learning_generators()
-
-
-# Export all generators
-__all__ = [
-    'EnhancedRepresentationLearningGenerator',
-    'PatchTSTRepresentationGenerator',
-    'TFTEncoderRepresentationGenerator',
-    'create_enhanced_representation_learning_generators',
-    'create_default_enhanced_representation_learning_generators'
-]
